+++ dropwizard-db/src/main/java/com/codahale/dropwizard/db/ConfigurationStrategy.java
@@ -1,7 +1,0 @@
-package com.codahale.dropwizard.db;
-
-import com.codahale.dropwizard.Configuration;
-
-public interface ConfigurationStrategy<T extends Configuration> {
-    DataSourceFactory getDatabaseFactory(T configuration);
-}