+++ dropwizard-db/src/main/java/com/codahale/dropwizard/db/DatabaseConfiguration.java
@@ -1,0 +1,7 @@
+package com.codahale.dropwizard.db;
+
+import com.codahale.dropwizard.Configuration;
+
+public interface DatabaseConfiguration<T extends Configuration> {
+    DataSourceFactory getDataSourceFactory(T configuration);
+}